tallyWhatIDontKnow

+++ src/Expense.java
@@ -2,9 +2,9 @@ import java.util.HashMap;
 
 public interface Expense
 {
-    default public HashMap<Destination, Integer> getCost()
+    default public HashMap<Destination, Float> getCost()
     {
-        HashMap<Destination, Integer> vacationMap = new HashMap<Destination, Integer>();
+        HashMap<Destination, Float> vacationMap = new HashMap<Destination, Float>();
         return vacationMap;
     }
